1
1

+++ backend/app/models/notification.py
@@ -1,7 +1,7 @@
 """通知模型"""
 from datetime import datetime
 
-from sqlalchemy import BigInteger, Boolean, DateTime, SmallInteger, String, func
+from sqlalchemy import Integer, Boolean, DateTime, SmallInteger, String, func
 from sqlalchemy.orm import Mapped, mapped_column
 
 from ..database import Base
@@ -10,11 +10,11 @@ from ..database import Base
 class Notification(Base):
     __tablename__ = "notifications"
 
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, index=True)
-    user_id: Mapped[int] = mapped_column(BigInteger, index=True)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, index=True)
+    user_id: Mapped[int] = mapped_column(Integer, index=True)
     type: Mapped[int] = mapped_column(SmallInteger)  # 1入队申请 2入队邀请 3离队通知 4系统通知
     content: Mapped[str] = mapped_column(String(500))
     related_type: Mapped[str | None] = mapped_column(String(30))  # team/request/invite
-    related_id: Mapped[int | None] = mapped_column(BigInteger)
+    related_id: Mapped[int | None] = mapped_column(Integer)
     is_read: Mapped[bool] = mapped_column(Boolean, default=False)
     created_at: Mapped[datetime] = mapped_column(DateTime, server_default=func.now())
